+++ src/main/resources/application.yml
@@ -1,9 +1,9 @@
 spring:
   profiles:
     active:
-      - local
+      - dev
     include:
-      - oauth2-local
-      - jwt-local
-      - storage-s3-local
+      - oauth2-dev
+      - jwt-dev
+      - storage-s3-dev
       - open-ai
